Move test support classes to their own package
- this improves readability by keeping actual tests and test support classes separate
