+++ acceptance-tests/src/test/java/me/atam/atam4jsampleapp/FailingTestAcceptanceTest.java
@@ -2,6 +2,8 @@ package me.atam.atam4jsampleapp;
 
 import me.atam.atam4j.dummytests.FailingTest;
 import me.atam.atam4j.health.AcceptanceTestsHealthCheck;
+import me.atam.atam4jsampleapp.testsupport.AcceptanceTest;
+import me.atam.atam4jsampleapp.testsupport.Atam4jApplicationStarter;
 import org.junit.Test;
 
 public class FailingTestAcceptanceTest extends AcceptanceTest {
