+++ acceptance-tests/src/test/java/me/atam/atam4jsampleapp/PassingTestAcceptanceTest.java
@@ -3,11 +3,14 @@ package me.atam.atam4jsampleapp;
 import me.atam.atam4j.PollingPredicate;
 import me.atam.atam4j.dummytests.PassingTest;
 import me.atam.atam4j.health.AcceptanceTestsHealthCheck;
+import me.atam.atam4jsampleapp.testsupport.AcceptanceTest;
+import me.atam.atam4jsampleapp.testsupport.Atam4jApplicationStarter;
+import me.atam.atam4jsampleapp.testsupport.HealthCheckResult;
 import org.junit.Test;
 
 import javax.ws.rs.core.Response;
 
-public class PassingTestAcceptanceTest extends AcceptanceTest{
+public class PassingTestAcceptanceTest extends AcceptanceTest {
 
     public static final int MAX_ATTEMPTS = 2000;
     public static final int RETRY_POLL_INTERVAL = 1;
