+++ acceptance-tests/src/test/java/me/atam/atam4jsampleapp/testsupport/HealthCheckResult.java
@@ -1,4 +1,4 @@
-package me.atam.atam4jsampleapp;
+package me.atam.atam4jsampleapp.testsupport;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
